# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentHub Multi-Agent Chatbot Framework",
    "description": "Production-ready multi-tenant chatbot framework using LangChain 0.3+",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/{tenant_id}/chat": {
      "post": {
        "tags": [
          "chat",
          "chat"
        ],
        "summary": "Chat Endpoint",
        "description": "Process user message and return agent response.\n\nFlow:\n1. Validate tenant access\n2. Create or retrieve session\n3. Save user message\n4. Route to SupervisorAgent\n5. Save assistant response\n6. Track performance metrics",
        "operationId": "chat_endpoint_api__tenant_id__chat_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "description": "User message content"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Existing session ID for follow-up messages"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "description": "Request schema for chat endpoint."
      },
      "ChatResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Response status: success/error/clarification_needed"
          },
          "agent": {
            "type": "string",
            "title": "Agent",
            "description": "Agent that processed the request"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "description": "Detected user intent"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Response data (structure varies by agent)"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "description": "Output format type"
          },
          "renderer_hint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Renderer Hint",
            "description": "UI rendering hints"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Response metadata"
          }
        },
        "type": "object",
        "required": [
          "status",
          "agent",
          "intent",
          "data",
          "format",
          "metadata"
        ],
        "title": "ChatResponse",
        "description": "Response schema for chat endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}